refactor: extract daemon cli shared helpers
